Added toolbar and solved problem with mixing A and Z in local and web server scripts

+++ calculate_mass_excess.py
@@ -91,8 +91,8 @@ def calculate_v3(m1, m2, m3, E_x, T_1):
     m1 = float(m1)
     m2 = float(m2)
     m3 = float(m3)
-    E_x = int(E_x)
-    T_1 = int(T_1)
+    E_x = float(E_x)
+    T_1 = float(T_1)
     print("DATA: ", m1, m2, m3, E_x, T_1)
     T_i = (m2 / (m1 + m2)) * T_1
     Q = (m1 + m2) - (m3 + m1)
